+++ include/tasks/Quiz.hpp
@@ -3,8 +3,9 @@
 
 // Task header file
 #include "tasks/GradedTask.hpp"
+
+// Built in library
 #include <string>
-#include <cstdint>
 
 
 class Quiz : public GradedTask {
@@ -17,7 +18,7 @@ class Quiz : public GradedTask {
             const std::string& name, 
             const std::string& date, 
             bool status,
-            uint16_t grade
+            int grade
         );
         // Accessor
         std::string getType();
